started with detector visulization example
